refactor(graph): decompose sqlite_graph.go via /evolve [evolve-simplify]
First /evolve-driven decomposition. sqlite_graph.go 1598 → 547 LOC across 5 files. API preserved (35 methods + ctor + type unchanged). Coverage 75.2% baseline maintained. Race -count=2 clean. Skeptic-evaluator verdict PASS (verified 110-line gate flag is dissimilarity-index false positive, filed mache-301daf). Closes mache-13f80f, mache-140be5, mache-141c81, mache-143454.
