+++ internal/graph/sqlite_graph.go
@@ -1,17 +1,12 @@
 package graph
 
 import (
-	"bytes"
-	"context"
 	"database/sql"
 	"encoding/json"
 	"fmt"
 	"io/fs"
-	"log"
 	"os"
 	"path/filepath"
-	"regexp"
-	"sort"
 	"strings"
 	"sync"
 	"time"
@@ -104,17 +99,6 @@ type SQLiteGraph struct {
 
 	extractor CallExtractor
 	defs      map[string][]string // symbol_name → []construct_dir_id (populated by AddDef)
-}
-
-// schemaLevel is a compiled representation of one level in the schema tree.
-type schemaLevel struct {
-	nameRaw    string
-	selector   string
-	isStatic   bool
-	staticName string
-	children   []*schemaLevel
-	files      []api.Leaf
-	depth      int
 }
 
 // EagerScan pre-scans all root nodes so no FUSE callback ever blocks on a scan.
@@ -250,31 +234,6 @@ func OpenSQLiteGraph(dbPath string, schema *api.Topology, render TemplateRendere
 	}, nil
 }
 
-func compileLevels(schema *api.Topology) []*schemaLevel {
-	var out []*schemaLevel
-	for _, node := range schema.Nodes {
-		out = append(out, compileOneLevel(node, 0))
-	}
-	return out
-}
-
-func compileOneLevel(node api.Node, depth int) *schemaLevel {
-	l := &schemaLevel{
-		nameRaw:  node.Name,
-		selector: node.Selector,
-		files:    node.Files,
-		depth:    depth,
-	}
-	if !strings.Contains(node.Name, "{{") {
-		l.isStatic = true
-		l.staticName = node.Name
-	}
-	for _, child := range node.Children {
-		l.children = append(l.children, compileOneLevel(child, depth+1))
-	}
-	return l
-}
-
 // ---------------------------------------------------------------------------
 // Graph interface
 // ---------------------------------------------------------------------------
@@ -290,160 +249,6 @@ func (g *SQLiteGraph) SetCallExtractor(fn CallExtractor) {
 // to this .db (mache-190508 step 3).
 func (g *SQLiteGraph) DBPath() string {
 	return g.dbPath
-}
-
-// AddDef records that a construct (dirID) defines the given token.
-func (g *SQLiteGraph) AddDef(token, dirID string) error {
-	g.pendingMu.Lock()
-	defer g.pendingMu.Unlock()
-	if g.defs == nil {
-		g.defs = make(map[string][]string)
-	}
-	g.defs[token] = append(g.defs[token], dirID)
-	return nil
-}
-
-// DefsMap returns a snapshot of the token→dirIDs definition map.
-func (g *SQLiteGraph) DefsMap() map[string][]string {
-	g.pendingMu.Lock()
-	defer g.pendingMu.Unlock()
-	cp := make(map[string][]string, len(g.defs))
-	for k, v := range g.defs {
-		cp[k] = append([]string(nil), v...)
-	}
-	return cp
-}
-
-// SearchDefs returns up to `limit` token→nodeIDs entries whose
-// token matches the SQL LIKE pattern. Used by `search role=definition`
-// to avoid materializing the full defs map.
-//
-// For pre-built .db files the in-memory g.defs map is empty; this
-// pushes the filter down to SQL against node_defs. Mirrors LookupDef's
-// fallback structure: in-memory first (with sqlLikeMatch-style check),
-// then SQL. Bead mache-9cba08.
-func (g *SQLiteGraph) SearchDefs(pattern string, limit int) map[string][]string {
-	if limit <= 0 {
-		limit = 100
-	}
-	out := make(map[string][]string)
-
-	g.pendingMu.Lock()
-	for token, ids := range g.defs {
-		if !likeMatch(pattern, token) {
-			continue
-		}
-		cp := make([]string, len(ids))
-		copy(cp, ids)
-		out[token] = cp
-		if len(out) >= limit {
-			g.pendingMu.Unlock()
-			return out
-		}
-	}
-	g.pendingMu.Unlock()
-
-	if !g.useNodesTable || g.db == nil {
-		return out
-	}
-	rows, err := g.db.Query(
-		"SELECT token, node_id FROM node_defs WHERE token LIKE ? LIMIT ?",
-		pattern, limit-len(out),
-	)
-	if err != nil {
-		return out
-	}
-	defer func() { _ = rows.Close() }()
-	for rows.Next() {
-		var token, nodeID string
-		if scanErr := rows.Scan(&token, &nodeID); scanErr != nil {
-			continue
-		}
-		// In-memory had precedence — don't override an existing entry.
-		if _, exists := out[token]; exists {
-			continue
-		}
-		out[token] = append(out[token], nodeID)
-		if len(out) >= limit {
-			return out
-		}
-	}
-	return out
-}
-
-// likeMatch is the in-memory equivalent of SQL LIKE for SearchDefs's
-// in-memory pass. Supports % (any-chars) and _ (single-char). Kept
-// in-package to avoid pulling the cmd/-level helper across boundaries.
-func likeMatch(pattern, value string) bool {
-	// Empty pattern matches nothing (consistent with sqlLikeMatch).
-	if pattern == "" {
-		return false
-	}
-	return likeMatchRec(pattern, value)
-}
-
-func likeMatchRec(pattern, value string) bool {
-	for i := 0; i < len(pattern); i++ {
-		switch pattern[i] {
-		case '%':
-			rest := pattern[i+1:]
-			if rest == "" {
-				return true
-			}
-			for j := 0; j <= len(value); j++ {
-				if likeMatchRec(rest, value[j:]) {
-					return true
-				}
-			}
-			return false
-		case '_':
-			if len(value) == 0 {
-				return false
-			}
-			value = value[1:]
-		default:
-			if len(value) == 0 || value[0] != pattern[i] {
-				return false
-			}
-			value = value[1:]
-		}
-	}
-	return len(value) == 0
-}
-
-// LookupDef returns the dir IDs that define `token` without
-// snapshotting the entire defs map. Returns nil for unknown tokens.
-// Mirrors MemoryStore.LookupDef; see that method's docstring.
-//
-// For pre-built DBs (useNodesTable path) the in-memory g.defs map is
-// usually empty because data lives in the `node_defs` SQL table. Falls
-// back to the SQL query on in-memory miss so find_definition works
-// against ley-line-parsed databases. Matches the SQL pattern used by
-// GetCallees (line ~722 / ~761) for parity.
-func (g *SQLiteGraph) LookupDef(token string) []string {
-	g.pendingMu.Lock()
-	ids, ok := g.defs[token]
-	g.pendingMu.Unlock()
-	if ok {
-		return append([]string(nil), ids...)
-	}
-
-	if !g.useNodesTable || g.db == nil {
-		return nil
-	}
-	rows, err := g.db.Query("SELECT node_id FROM node_defs WHERE token = ?", token)
-	if err != nil {
-		return nil
-	}
-	defer func() { _ = rows.Close() }()
-	var dirIDs []string
-	for rows.Next() {
-		var dirID string
-		if scanErr := rows.Scan(&dirID); scanErr == nil && dirID != "" {
-			dirIDs = append(dirIDs, dirID)
-		}
-	}
-	return dirIDs
 }
 
 func (g *SQLiteGraph) GetNode(id string) (*Node, error) {
@@ -632,363 +437,6 @@ func (g *SQLiteGraph) ReadContent(id string, buf []byte, offset int64) (int, err
 	return SliceContent(content, buf, offset), nil
 }
 
-// AddRef accumulates a reference in-memory. No SQL is issued until FlushRefs.
-// This eliminates the read-modify-write cycle per call — all bitmap mutations
-// happen in RAM, and FlushRefs writes them in a single transaction.
-// Not used for nodes-table path (refs already in main DB from mache build).
-func (g *SQLiteGraph) AddRef(token, nodeID string) error {
-	if g.useNodesTable {
-		return nil // refs already in main DB
-	}
-	g.pendingMu.Lock()
-	defer g.pendingMu.Unlock()
-
-	fid, ok := g.fileIDMap[nodeID]
-	if !ok {
-		fid = g.nextFileID
-		g.nextFileID++
-		g.fileIDMap[nodeID] = fid
-	}
-
-	bm, ok := g.pendingRefs[token]
-	if !ok {
-		bm = roaring.New()
-		g.pendingRefs[token] = bm
-	}
-	bm.Add(fid)
-	return nil
-}
-
-// FlushRefs writes all accumulated refs to the sidecar database in a single
-// transaction. Call once after ingestion is complete. Batches all inserts
-// (len(fileIDMap) + len(pendingRefs)) into one transaction.
-//
-// Guarded by sync.Once — safe to call multiple times; only the first call
-// performs the flush. This prevents the double-call bug where a second flush
-// would reset nextFileID to 0, causing ID collisions in file_ids.
-func (g *SQLiteGraph) FlushRefs() error {
-	var flushErr error
-	g.flushOnce.Do(func() {
-		flushErr = g.flushRefsInternal()
-	})
-	return flushErr
-}
-
-func (g *SQLiteGraph) flushRefsInternal() error {
-	g.pendingMu.Lock()
-	refs := g.pendingRefs
-	fileIDs := g.fileIDMap
-	g.pendingMu.Unlock()
-
-	if len(refs) == 0 {
-		return nil
-	}
-
-	tx, err := g.refsDB.Begin()
-	if err != nil {
-		return fmt.Errorf("begin refs flush: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }() // safe to ignore
-
-	// Write file_ids
-	fileStmt, err := tx.Prepare("INSERT OR IGNORE INTO file_ids (id, path) VALUES (?, ?)")
-	if err != nil {
-		return fmt.Errorf("prepare file_ids insert: %w", err)
-	}
-	defer func() { _ = fileStmt.Close() }() // safe to ignore
-
-	for path, id := range fileIDs {
-		if _, err := fileStmt.Exec(id, path); err != nil {
-			return fmt.Errorf("insert file_id %s: %w", path, err)
-		}
-	}
-
-	// Write bitmaps
-	refStmt, err := tx.Prepare("INSERT OR REPLACE INTO node_refs (token, bitmap) VALUES (?, ?)")
-	if err != nil {
-		return fmt.Errorf("prepare node_refs insert: %w", err)
-	}
-	defer func() { _ = refStmt.Close() }() // safe to ignore
-
-	var buf bytes.Buffer
-	for token, bm := range refs {
-		buf.Reset()
-		if _, err := bm.WriteTo(&buf); err != nil {
-			return fmt.Errorf("serialize bitmap for %s: %w", token, err)
-		}
-		if _, err := refStmt.Exec(token, buf.Bytes()); err != nil {
-			return fmt.Errorf("insert ref %s: %w", token, err)
-		}
-	}
-
-	return tx.Commit()
-}
-
-// GetCallers returns the list of files (nodes) that reference the given token.
-// For nodes-table path: queries main DB's node_refs (token, node_id) directly.
-// For legacy path: reads roaring bitmaps from the sidecar refs database.
-func (g *SQLiteGraph) GetCallers(token string) ([]*Node, error) {
-	if g.useNodesTable {
-		return g.getCallersFromMainDB(token)
-	}
-	return g.getCallersFromSidecar(token)
-}
-
-// GetCallees implements Graph.
-func (g *SQLiteGraph) GetCallees(id string) ([]*Node, error) {
-	id = NormalizeID(id)
-
-	// 1. Find the "source" file child
-	children, err := g.ListChildren(id)
-	if err != nil {
-		return nil, nil
-	}
-
-	var sourceID string
-	for _, child := range children {
-		base := filepath.Base(child)
-		if base == "source" {
-			// Both backends now return full paths from ListChildren —
-			// SQLiteGraph forwards to NodesTableReader which selects the
-			// `id` column (full path); the sidecar path also stores full
-			// paths in dirChildren. Earlier this branch concatenated
-			// `id + "/" + child` for useNodesTable assuming bare names —
-			// produced "leyline/methods/SendOp/leyline/methods/SendOp/source"
-			// and made find_callees silently return [] for every construct
-			// on a pre-built .db. Bead mache-9ca6af root cause.
-			sourceID = child
-			break
-		}
-	}
-	if sourceID == "" {
-		return nil, nil
-	}
-
-	// 2. Read content
-	segments := strings.Split(sourceID, "/")
-	_, fileLeaf := g.walkSchema(segments)
-
-	if fileLeaf == nil && !g.useNodesTable {
-		return nil, nil
-	}
-
-	content, err := g.resolveContent(sourceID, segments, fileLeaf)
-	if err != nil {
-		return nil, nil
-	}
-
-	// 3. Determine langName from construct node Properties (stored in record column)
-	var langName string
-	if g.useNodesTable {
-		var recordJSON sql.NullString
-		// kind = 1 → graph.NodeKindDir (constructs are dirs).
-		_ = g.db.QueryRow("SELECT record FROM nodes WHERE id = ? AND kind = 1", id).Scan(&recordJSON)
-		if recordJSON.Valid && recordJSON.String != "" {
-			var props map[string][]byte
-			if json.Unmarshal([]byte(recordJSON.String), &props) == nil {
-				if lang, ok := props["lang"]; ok {
-					langName = string(lang)
-				}
-			}
-		}
-	}
-
-	// 4. Extract qualified calls
-	if g.extractor == nil {
-		return nil, nil
-	}
-	qcalls, err := g.extractor(content, sourceID, langName)
-	if err != nil {
-		return nil, fmt.Errorf("extract calls: %w", err)
-	}
-	if len(qcalls) == 0 {
-		return nil, nil
-	}
-
-	// 5. Resolve tokens to definition nodes (qualified → bare → SQL fallback)
-	var nodes []*Node
-	seen := make(map[string]bool)
-
-	// Snapshot defs under lock — deep copy slices to prevent races with concurrent AddDef.
-	var defs map[string][]string
-	g.pendingMu.Lock()
-	if g.defs != nil {
-		defs = make(map[string][]string, len(g.defs))
-		for k, v := range g.defs {
-			cp := make([]string, len(v))
-			copy(cp, v)
-			defs[k] = cp
-		}
-	}
-	g.pendingMu.Unlock()
-
-	for _, qc := range qcalls {
-		resolved := false
-
-		// Qualified resolution: "auth.Validate" → defs["auth.Validate"]
-		if qc.Qualifier != "" {
-			qualKey := qc.Qualifier + "." + qc.Token
-			if defs != nil {
-				if defIDs, ok := defs[qualKey]; ok {
-					for _, defID := range defIDs {
-						if defID == id || seen[defID] {
-							continue
-						}
-						seen[defID] = true
-						nodes = append(nodes, &Node{ID: defID, Mode: os.ModeDir | 0o555})
-						resolved = true
-					}
-				}
-			}
-			// SQL fallback: node_defs table (pre-built DBs)
-			if !resolved && g.useNodesTable {
-				rows, qErr := g.db.Query("SELECT node_id FROM node_defs WHERE token = ?", qualKey)
-				if qErr == nil {
-					for rows.Next() {
-						var defID string
-						if rows.Scan(&defID) == nil && defID != id && !seen[defID] {
-							seen[defID] = true
-							nodes = append(nodes, &Node{ID: defID, Mode: os.ModeDir | 0o555})
-							resolved = true
-						}
-					}
-					_ = rows.Close()
-				}
-			}
-			if resolved {
-				continue
-			}
-		}
-
-		// Bare token lookup
-		if defs != nil {
-			if defIDs, ok := defs[qc.Token]; ok {
-				for _, defID := range defIDs {
-					if defID == id || seen[defID] {
-						continue
-					}
-					seen[defID] = true
-					nodes = append(nodes, &Node{ID: defID, Mode: os.ModeDir | 0o555})
-					resolved = true
-				}
-				if resolved {
-					continue
-				}
-			}
-		}
-
-		// SQL fallback: node_defs then nodes table
-		if g.useNodesTable {
-			var defID string
-			err := g.db.QueryRow("SELECT node_id FROM node_defs WHERE token = ? LIMIT 1", qc.Token).Scan(&defID)
-			if err == nil && defID != id && !seen[defID] {
-				seen[defID] = true
-				nodes = append(nodes, &Node{ID: defID, Mode: os.ModeDir | 0o555})
-				continue
-			}
-
-			// Suffix-match fallback for receiver-method calls (bead
-			// mache-9ca6af). A call like `c.sendRaw(...)` extracts as
-			// the bare token "sendRaw"; node_defs has it only as the
-			// qualified form "SocketClient.sendRaw". Bare lookup above
-			// misses. If there's exactly ONE "*.token" entry in
-			// node_defs, that's our target — unambiguous resolution
-			// for the receiver case. >1 matches means ambiguous (two
-			// types both define a method of this name) — skip rather
-			// than guess wrongly. LIMIT 2 detects ambiguity cheaply.
-			rows, qErr := g.db.Query(
-				"SELECT node_id FROM node_defs WHERE token LIKE ? LIMIT 2",
-				"%."+qc.Token,
-			)
-			if qErr == nil {
-				var candidates []string
-				for rows.Next() {
-					var nodeID string
-					if scanErr := rows.Scan(&nodeID); scanErr == nil && nodeID != id {
-						candidates = append(candidates, nodeID)
-					}
-				}
-				_ = rows.Close()
-				if len(candidates) == 1 && !seen[candidates[0]] {
-					seen[candidates[0]] = true
-					nodes = append(nodes, &Node{ID: candidates[0], Mode: os.ModeDir | 0o555})
-					continue
-				}
-			}
-
-			// Final fallback: name match in nodes table
-			// kind = 1 → graph.NodeKindDir (definition constructs are dirs).
-			err = g.db.QueryRow("SELECT id FROM nodes WHERE name = ? AND kind = 1 LIMIT 1", qc.Token).Scan(&defID)
-			if err == nil && defID != id && !seen[defID] {
-				seen[defID] = true
-				nodes = append(nodes, &Node{ID: defID, Mode: os.ModeDir | 0o555})
-			}
-		}
-	}
-
-	return nodes, nil
-}
-
-// getCallersFromMainDB queries the main DB's node_refs table directly.
-// node_refs schema: (token TEXT, node_id TEXT) — written by mache build.
-func (g *SQLiteGraph) getCallersFromMainDB(token string) ([]*Node, error) {
-	return g.ntr.GetCallers(token)
-}
-
-// getCallersFromSidecar reads roaring bitmaps from the sidecar .refs.db.
-func (g *SQLiteGraph) getCallersFromSidecar(token string) ([]*Node, error) {
-	var blob []byte
-	err := g.refsDB.QueryRow("SELECT bitmap FROM node_refs WHERE token = ?", token).Scan(&blob)
-	if err == sql.ErrNoRows {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	rb := roaring.New()
-	if err := rb.UnmarshalBinary(blob); err != nil {
-		return nil, fmt.Errorf("unmarshal bitmap: %w", err)
-	}
-
-	var fileIDs []uint32
-	it := rb.Iterator()
-	for it.HasNext() {
-		fileIDs = append(fileIDs, it.Next())
-	}
-	if len(fileIDs) == 0 {
-		return nil, nil
-	}
-
-	args := make([]any, len(fileIDs))
-	placeholders := make([]string, len(fileIDs))
-	for i, id := range fileIDs {
-		args[i] = id
-		placeholders[i] = "?"
-	}
-
-	query := fmt.Sprintf("SELECT path FROM file_ids WHERE id IN (%s)", strings.Join(placeholders, ","))
-	rows, err := g.refsDB.Query(query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("query file paths: %w", err)
-	}
-	defer func() { _ = rows.Close() }() // safe to ignore
-
-	var nodes []*Node
-	for rows.Next() {
-		var path string
-		if err := rows.Scan(&path); err != nil {
-			log.Printf("GetCallers: skip row scan: %v", err)
-			continue
-		}
-		nodes = append(nodes, &Node{
-			ID:   path,
-			Mode: 0o444,
-		})
-	}
-	return nodes, nil
-}
-
 // Invalidate evicts cached size and content for a node.
 // Must be called after write-back modifies a file's content to prevent
 // stale size/data from being served on the next Getattr or Read.
@@ -1011,92 +459,6 @@ func (g *SQLiteGraph) DB() *sql.DB {
 	return g.db
 }
 
-// QueryRefs executes a SQL query against the refs database.
-// For nodes-table path: queries the main DB (node_refs has (token, node_id)).
-// For legacy path: queries the sidecar (includes mache_refs virtual table).
-func (g *SQLiteGraph) QueryRefs(query string, args ...any) (*sql.Rows, error) {
-	if g.useNodesTable {
-		return g.db.Query(query, args...)
-	}
-	return g.refsDB.Query(query, args...)
-}
-
-// RefsMap returns a token→nodeIDs map for community detection.
-// For the nodes-table path, queries node_refs (token, node_id).
-// For the legacy bitmap path, decodes bitmaps and resolves file IDs.
-func (g *SQLiteGraph) RefsMap() map[string][]string {
-	if g.useNodesTable {
-		return g.refsMapFromNodesTable()
-	}
-	return g.refsMapFromBitmaps()
-}
-
-func (g *SQLiteGraph) refsMapFromNodesTable() map[string][]string {
-	rows, err := g.db.Query("SELECT token, node_id FROM node_refs")
-	if err != nil {
-		return nil
-	}
-	defer func() { _ = rows.Close() }()
-
-	refs := map[string][]string{}
-	for rows.Next() {
-		var token, nodeID string
-		if err := rows.Scan(&token, &nodeID); err != nil {
-			continue
-		}
-		refs[token] = append(refs[token], nodeID)
-	}
-	return refs
-}
-
-func (g *SQLiteGraph) refsMapFromBitmaps() map[string][]string {
-	if g.refsDB == nil {
-		return nil
-	}
-
-	// Build reverse file ID lookup
-	fileRows, err := g.refsDB.Query("SELECT id, path FROM file_ids")
-	if err != nil {
-		return nil
-	}
-	idToPath := map[uint32]string{}
-	for fileRows.Next() {
-		var id uint32
-		var path string
-		if err := fileRows.Scan(&id, &path); err != nil {
-			continue
-		}
-		idToPath[id] = path
-	}
-	_ = fileRows.Close()
-
-	// Read bitmaps
-	refRows, err := g.refsDB.Query("SELECT token, bitmap FROM node_refs")
-	if err != nil {
-		return nil
-	}
-	defer func() { _ = refRows.Close() }()
-
-	refs := map[string][]string{}
-	for refRows.Next() {
-		var token string
-		var blob []byte
-		if err := refRows.Scan(&token, &blob); err != nil {
-			continue
-		}
-		bm := roaring.New()
-		if _, err := bm.ReadFrom(bytes.NewReader(blob)); err != nil {
-			continue
-		}
-		for _, id := range bm.ToArray() {
-			if path, ok := idToPath[id]; ok {
-				refs[token] = append(refs[token], path)
-			}
-		}
-	}
-	return refs
-}
-
 // Close closes both the source and sidecar database connections.
 func (g *SQLiteGraph) Close() error {
 	// Unregister from vtab module to prevent leaks/races (sidecar path only)
@@ -1115,423 +477,10 @@ func (g *SQLiteGraph) Close() error {
 	return err
 }
 
-// ---------------------------------------------------------------------------
-// Schema walking
-// ---------------------------------------------------------------------------
-
 // walkSchema maps a path to its schema level and (if a file) leaf definition.
 // Returns (level, nil) for directories, (level, &leaf) for files, (nil, nil) for invalid paths.
 func (g *SQLiteGraph) walkSchema(segments []string) (*schemaLevel, *api.Leaf) {
 	return walkSchemaLevels(g.levels, segments)
-}
-
-// walkSchemaLevels walks compiled schema levels to find the level and optional
-// leaf matching the given path segments. Shared by SQLiteGraph and WritableGraph.
-//
-// When descending past a level with multiple children, prefer the static
-// child whose name matches the path segment (e.g. go-schema's "functions" /
-// "methods" / "types" siblings under each package). Fall back to the first
-// dynamic (template-named) child when no static name matches.
-func walkSchemaLevels(levels []*schemaLevel, segments []string) (*schemaLevel, *api.Leaf) {
-	if len(segments) == 0 {
-		return nil, nil
-	}
-
-	var root *schemaLevel
-	for _, l := range levels {
-		if l.isStatic && l.staticName == segments[0] {
-			root = l
-			break
-		}
-	}
-	if root == nil {
-		return nil, nil
-	}
-	if len(segments) == 1 {
-		return root, nil
-	}
-
-	current := root
-	for i := 1; i < len(segments); i++ {
-		seg := segments[i]
-
-		// Check if this segment matches a file at the current level
-		for j := range current.files {
-			fname := current.files[j].Name
-			if !strings.Contains(fname, "{{") && fname == seg {
-				return current, &current.files[j]
-			}
-		}
-
-		if len(current.children) == 0 {
-			return nil, nil
-		}
-		current = pickChildLevel(current.children, seg)
-		if current == nil {
-			return nil, nil
-		}
-	}
-
-	return current, nil
-}
-
-// pickChildLevel selects the schema child that matches the given path segment.
-// Returns the matching static child if one exists, otherwise the first dynamic
-// (template-named) child, otherwise the first child as a last-resort fallback.
-// Returns nil only when the input slice is empty.
-func pickChildLevel(children []*schemaLevel, seg string) *schemaLevel {
-	if len(children) == 0 {
-		return nil
-	}
-	for _, c := range children {
-		if c.isStatic && c.staticName == seg {
-			return c
-		}
-	}
-	for _, c := range children {
-		if !c.isStatic {
-			return c
-		}
-	}
-	return children[0]
-}
-
-// ---------------------------------------------------------------------------
-// Lazy scanning
-// ---------------------------------------------------------------------------
-
-func (g *SQLiteGraph) ensureScanned(rootName string) error {
-	val, _ := g.scanOnce.LoadOrStore(rootName, &sync.Once{})
-	var err error
-	val.(*sync.Once).Do(func() {
-		err = g.scanRoot(rootName)
-		if err != nil {
-			g.scanErr.Store(rootName, err)
-		}
-	})
-	if err != nil {
-		return err
-	}
-	if v, ok := g.scanErr.Load(rootName); ok {
-		return v.(error)
-	}
-	return nil
-}
-
-// --- Scan types ---
-
-type scanResult struct {
-	entries  []pathEntry
-	leafDirs []leafMapping
-}
-
-type pathEntry struct {
-	parent string
-	child  string
-}
-
-type leafMapping struct {
-	dirPath  string
-	recordID string
-}
-
-// --- Field extraction from name templates ---
-
-// fieldRefRe matches Go template field references like .item.cve.id
-var fieldRefRe = regexp.MustCompile(`\.(\w+(?:\.\w+)*)`)
-
-// collectNameTemplates gathers all dynamic name template strings from the schema tree.
-func collectNameTemplates(level *schemaLevel) []string {
-	var tmpls []string
-	var walk func(*schemaLevel)
-	walk = func(l *schemaLevel) {
-		if !l.isStatic {
-			tmpls = append(tmpls, l.nameRaw)
-		}
-		for _, c := range l.children {
-			walk(c)
-		}
-	}
-	walk(level)
-	return tmpls
-}
-
-// extractFieldPaths pulls dotted field references from Go templates.
-// e.g. "{{slice .item.cve.id 4 8}}" → ["item.cve.id"]
-func extractFieldPaths(templates []string) []string {
-	seen := make(map[string]bool)
-	for _, tmpl := range templates {
-		for _, m := range fieldRefRe.FindAllStringSubmatch(tmpl, -1) {
-			seen[m[1]] = true
-		}
-	}
-	paths := make([]string, 0, len(seen))
-	for p := range seen {
-		paths = append(paths, p)
-	}
-	sort.Strings(paths)
-	return paths
-}
-
-// buildScanQuery builds a SELECT using json_extract for only the fields
-// needed by name templates. Avoids transferring and parsing full record JSON.
-func buildScanQuery(fieldPaths []string, tableName string) string {
-	cols := make([]string, 0, len(fieldPaths)+1)
-	cols = append(cols, "id")
-	for _, fp := range fieldPaths {
-		cols = append(cols, fmt.Sprintf("json_extract(record, '$.%s')", fp))
-	}
-	return "SELECT " + strings.Join(cols, ", ") + " FROM " + tableName
-}
-
-// setNestedField builds a nested map from a dotted path.
-// e.g. setNestedField(m, "item.cve.id", "CVE-2024-0001")
-//
-//	→ m["item"]["cve"]["id"] = "CVE-2024-0001"
-func setNestedField(m map[string]any, dottedPath, value string) {
-	parts := strings.Split(dottedPath, ".")
-	current := m
-	for i, part := range parts {
-		if i == len(parts)-1 {
-			current[part] = value
-		} else {
-			if v, ok := current[part]; ok {
-				if nested, isMap := v.(map[string]any); isMap {
-					current = nested
-				} else {
-					// Path conflict: intermediate value is not a map; overwrite it
-					next := make(map[string]any)
-					current[part] = next
-					current = next
-				}
-			} else {
-				next := make(map[string]any)
-				current[part] = next
-				current = next
-			}
-		}
-	}
-}
-
-// --- Scan implementation ---
-
-// flushBatchSize is the number of records between batch flushes to sync.Map.
-// Keeps transient working-map memory bounded for large cross-reference scans.
-const flushBatchSize = 50000
-
-// scanRoot performs a single-pass streaming scan of all DB records to build the
-// directory tree for one root node. Uses json_extract to push field extraction
-// into SQLite, avoiding Go-side JSON parsing of the full record blob.
-//
-// Safety properties:
-//   - Read-only transaction: snapshot isolation prevents partial results if the
-//     source DB is being written to concurrently.
-//   - Batch flush: every flushBatchSize records, accumulated slices are sorted,
-//     deduped, and merged into sync.Map, then the working map is cleared. This
-//     bounds transient memory for 10M+ node cross-reference scans.
-//   - Error counting: scan/render failures are counted and logged rather than
-//     silently swallowed, so data drops are visible.
-//
-// Why single-threaded: The previous worker-pool implementation (NumCPU goroutines +
-// channels) was designed for CPU-bound template rendering, but profiling showed the
-// bottleneck is SQLite I/O, not rendering. Name templates are simple field lookups
-// (e.g. "{{.item.cve.id}}") that render in <1μs. The channel/goroutine overhead
-// actually hurt throughput and introduced deadlock risk. If a future schema uses
-// expensive template functions (regex, crypto), re-add parallelism — but measure first.
-func (g *SQLiteGraph) scanRoot(rootName string) error {
-	level := g.findRootLevel(rootName)
-	if level == nil {
-		return fmt.Errorf("root %q not found in schema", rootName)
-	}
-
-	// Analyze schema to find which fields the name templates need
-	fieldPaths := extractFieldPaths(collectNameTemplates(level))
-	query := buildScanQuery(fieldPaths, g.tableName)
-
-	// Read-only transaction for snapshot consistency — if the source DB is
-	// being written to during scan, we get a consistent point-in-time view.
-	tx, err := g.db.BeginTx(context.Background(), &sql.TxOptions{ReadOnly: true})
-	if err != nil {
-		return fmt.Errorf("begin scan tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }() // safe to ignore
-
-	rows, err := tx.Query(query)
-	if err != nil {
-		return fmt.Errorf("scan query: %w", err)
-	}
-	defer func() { _ = rows.Close() }() // safe to ignore
-
-	// Accumulate children as slices directly (no intermediate bool sets).
-	// Flushed to sync.Map every flushBatchSize records to bound memory.
-	childSlices := make(map[string][]string)
-	recIDs := make(map[string]string)
-	childSlices[rootName] = nil // ensure root exists even if DB is empty
-
-	// Reusable per-row scan buffers — allocated once, reused every iteration
-	nCols := len(fieldPaths) + 1
-	scanVals := make([]sql.NullString, nCols)
-	scanPtrs := make([]any, nCols)
-	for i := range scanVals {
-		scanPtrs[i] = &scanVals[i]
-	}
-	fields := make([]string, len(fieldPaths))
-
-	// Reusable result buffer for collectPathEntries
-	var result scanResult
-
-	count := 0
-	scanErrs := 0
-	nullSkips := 0
-	for rows.Next() {
-		if err := rows.Scan(scanPtrs...); err != nil {
-			scanErrs++
-			continue
-		}
-
-		// Check for NULL fields (records missing required template values)
-		skip := false
-		for i := range fieldPaths {
-			if !scanVals[i+1].Valid {
-				skip = true
-				break
-			}
-			fields[i] = scanVals[i+1].String
-		}
-		if skip {
-			nullSkips++
-			continue
-		}
-
-		// Build minimal values map and render schema path tree
-		values := make(map[string]any)
-		for i, path := range fieldPaths {
-			setNestedField(values, path, fields[i])
-		}
-
-		result.entries = result.entries[:0]
-		result.leafDirs = result.leafDirs[:0]
-		g.collectPathEntries(level, values, rootName, scanVals[0].String, &result)
-
-		for _, e := range result.entries {
-			childSlices[e.parent] = append(childSlices[e.parent], e.child)
-		}
-		for _, l := range result.leafDirs {
-			recIDs[l.dirPath] = l.recordID
-		}
-
-		count++
-		if count%100000 == 0 {
-			log.Printf("Scanning %d records...", count)
-		}
-
-		// Batch flush: merge accumulated data into sync.Map to bound memory
-		if count%flushBatchSize == 0 {
-			flushChildSlices(childSlices, &g.dirChildren)
-			for path, id := range recIDs {
-				g.recordIDs.Store(path, id)
-			}
-			// Clear working maps but keep root entry
-			childSlices = make(map[string][]string)
-			childSlices[rootName] = nil
-			recIDs = make(map[string]string)
-		}
-	}
-	if count >= 100000 {
-		log.Printf("Scanned %d records.", count)
-	}
-
-	if err := rows.Err(); err != nil {
-		return fmt.Errorf("scan rows: %w", err)
-	}
-
-	// Log skipped rows so data drops are visible
-	if scanErrs > 0 || nullSkips > 0 {
-		log.Printf("scan %q: %d records processed, %d scan errors, %d null-skipped",
-			rootName, count, scanErrs, nullSkips)
-	}
-
-	// Final flush of remaining data
-	flushChildSlices(childSlices, &g.dirChildren)
-	for path, id := range recIDs {
-		g.recordIDs.Store(path, id)
-	}
-
-	return nil
-}
-
-// flushChildSlices sorts, deduplicates, and merges accumulated child slices
-// into the sync.Map. For keys already in the map (from a previous batch),
-// the new children are merged and re-deduped.
-func flushChildSlices(slices map[string][]string, target *sync.Map) {
-	for parent, children := range slices {
-		// Sort and compact this batch
-		sort.Strings(children)
-		j := 0
-		for i, c := range children {
-			if i == 0 || c != children[i-1] {
-				children[j] = c
-				j++
-			}
-		}
-		deduped := children[:j]
-
-		// Merge with any existing entries from prior batches
-		if existing, ok := target.Load(parent); ok {
-			prev := existing.([]string)
-			deduped = mergeSortedDedup(prev, deduped)
-		}
-
-		target.Store(parent, deduped)
-	}
-}
-
-// collectPathEntries walks the schema children for one record, producing
-// parent→child entries and leaf directory→recordID mappings.
-func (g *SQLiteGraph) collectPathEntries(level *schemaLevel, values map[string]any, parentPath, recordID string, result *scanResult) {
-	for _, child := range level.children {
-		name, err := g.render(child.nameRaw, values)
-		if err != nil || name == "" {
-			continue
-		}
-
-		childPath := parentPath + "/" + name
-		result.entries = append(result.entries, pathEntry{parent: parentPath, child: childPath})
-
-		// Recurse into deeper directory levels
-		if len(child.children) > 0 {
-			g.collectPathEntries(child, values, childPath, recordID, result)
-		}
-
-		// Leaf directory: add file children and record mapping
-		if len(child.files) > 0 {
-			result.leafDirs = append(result.leafDirs, leafMapping{dirPath: childPath, recordID: recordID})
-			for _, f := range child.files {
-				result.entries = append(result.entries, pathEntry{parent: childPath, child: childPath + "/" + f.Name})
-			}
-		}
-	}
-}
-
-func (g *SQLiteGraph) findRootLevel(name string) *schemaLevel {
-	for _, l := range g.levels {
-		if l.isStatic && l.staticName == name {
-			return l
-		}
-	}
-	return nil
-}
-
-// isChild checks whether childPath appears in the cached children of parentPath.
-func (g *SQLiteGraph) isChild(parentPath, childPath string) bool {
-	v, ok := g.dirChildren.Load(parentPath)
-	if !ok {
-		return false
-	}
-	// Binary search on sorted children
-	children := v.([]string)
-	i := sort.SearchStrings(children, childPath)
-	return i < len(children) && children[i] == childPath
 }
 
 // ---------------------------------------------------------------------------
